fix endpoint for non-global (#6752)
* fix endpoint for non-global

* add a regional test

* add a doc
