+++ google/dialogflow_cx_operation.go
@@ -1,17 +1,3 @@
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
 package google
 
 import (
@@ -24,6 +10,7 @@ type DialogflowCXOperationWaiter struct {
 	Config    *Config
 	UserAgent string
 	CommonOperationWaiter
+	Location string
 }
 
 func (w *DialogflowCXOperationWaiter) QueryOp() (interface{}, error) {
@@ -31,15 +18,16 @@ func (w *DialogflowCXOperationWaiter) QueryOp() (interface{}, error) {
 		return nil, fmt.Errorf("Cannot query operation, it's unset or nil.")
 	}
 	// Returns the proper get.
-	url := fmt.Sprintf("https://dialogflow.googleapis.com/v3/%s", w.CommonOperationWaiter.Op.Name)
+	url := fmt.Sprintf("https://%s-dialogflow.googleapis.com/v3/%s", w.Location, w.CommonOperationWaiter.Op.Name)
 
 	return sendRequest(w.Config, "GET", "", url, w.UserAgent, nil)
 }
 
-func createDialogflowCXWaiter(config *Config, op map[string]interface{}, activity, userAgent string) (*DialogflowCXOperationWaiter, error) {
+func createDialogflowCXWaiter(config *Config, op map[string]interface{}, activity, userAgent, location string) (*DialogflowCXOperationWaiter, error) {
 	w := &DialogflowCXOperationWaiter{
 		Config:    config,
 		UserAgent: userAgent,
+		Location:  location,
 	}
 	if err := w.CommonOperationWaiter.SetOp(op); err != nil {
 		return nil, err
@@ -48,8 +36,8 @@ func createDialogflowCXWaiter(config *Config, op map[string]interface{}, activit
 }
 
 // nolint: deadcode,unused
-func dialogflowCXOperationWaitTimeWithResponse(config *Config, op map[string]interface{}, response *map[string]interface{}, activity, userAgent string, timeout time.Duration) error {
-	w, err := createDialogflowCXWaiter(config, op, activity, userAgent)
+func dialogflowCXOperationWaitTimeWithResponse(config *Config, op map[string]interface{}, response *map[string]interface{}, activity, userAgent, location string, timeout time.Duration) error {
+	w, err := createDialogflowCXWaiter(config, op, activity, userAgent, location)
 	if err != nil {
 		return err
 	}
@@ -59,12 +47,12 @@ func dialogflowCXOperationWaitTimeWithResponse(config *Config, op map[string]int
 	return json.Unmarshal([]byte(w.CommonOperationWaiter.Op.Response), response)
 }
 
-func dialogflowCXOperationWaitTime(config *Config, op map[string]interface{}, activity, userAgent string, timeout time.Duration) error {
+func dialogflowCXOperationWaitTime(config *Config, op map[string]interface{}, activity, userAgent, location string, timeout time.Duration) error {
 	if val, ok := op["name"]; !ok || val == "" {
 		// This was a synchronous call - there is no operation to wait for.
 		return nil
 	}
-	w, err := createDialogflowCXWaiter(config, op, activity, userAgent)
+	w, err := createDialogflowCXWaiter(config, op, activity, userAgent, location)
 	if err != nil {
 		// If w is nil, the op was synchronous.
 		return err
