+++ google/resource_dialogflow_cx_environment.go
@@ -1,17 +1,3 @@
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
 package google
 
 import (
@@ -162,7 +148,7 @@ func resourceDialogflowCXEnvironmentCreate(d *schema.ResourceData, meta interfac
 	// identity fields and d.Id() before read
 	var opRes map[string]interface{}
 	err = dialogflowCXOperationWaitTimeWithResponse(
-		config, res, &opRes, "Creating Environment", userAgent,
+		config, res, &opRes, "Creating Environment", userAgent, location,
 		d.Timeout(schema.TimeoutCreate))
 	if err != nil {
 		// The resource didn't actually create
@@ -328,7 +314,7 @@ func resourceDialogflowCXEnvironmentUpdate(d *schema.ResourceData, meta interfac
 	}
 
 	err = dialogflowCXOperationWaitTime(
-		config, res, "Updating Environment", userAgent,
+		config, res, "Updating Environment", userAgent, location,
 		d.Timeout(schema.TimeoutUpdate))
 
 	if err != nil {
@@ -381,7 +367,7 @@ func resourceDialogflowCXEnvironmentDelete(d *schema.ResourceData, meta interfac
 	}
 
 	err = dialogflowCXOperationWaitTime(
-		config, res, "Deleting Environment", userAgent,
+		config, res, "Deleting Environment", userAgent, location,
 		d.Timeout(schema.TimeoutDelete))
 
 	if err != nil {
