+++ google/resource_dialogflow_cx_version.go
@@ -1,17 +1,3 @@
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
 package google
 
 import (
@@ -181,7 +167,7 @@ func resourceDialogflowCXVersionCreate(d *schema.ResourceData, meta interface{})
 	// identity fields and d.Id() before read
 	var opRes map[string]interface{}
 	err = dialogflowCXOperationWaitTimeWithResponse(
-		config, res, &opRes, "Creating Version", userAgent,
+		config, res, &opRes, "Creating Version", userAgent, location,
 		d.Timeout(schema.TimeoutCreate))
 	if err != nil {
 		// The resource didn't actually create
@@ -340,7 +326,7 @@ func resourceDialogflowCXVersionUpdate(d *schema.ResourceData, meta interface{})
 	}
 
 	err = dialogflowCXOperationWaitTime(
-		config, res, "Updating Version", userAgent,
+		config, res, "Updating Version", userAgent, location,
 		d.Timeout(schema.TimeoutUpdate))
 
 	if err != nil {
@@ -393,7 +379,7 @@ func resourceDialogflowCXVersionDelete(d *schema.ResourceData, meta interface{})
 	}
 
 	err = dialogflowCXOperationWaitTime(
-		config, res, "Deleting Version", userAgent,
+		config, res, "Deleting Version", userAgent, location,
 		d.Timeout(schema.TimeoutDelete))
 
 	if err != nil {
