+++ google/resource_dialogflowcx_version_test.go
@@ -1,9 +1,12 @@
 package google
 
 import (
+	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
 func TestAccDialogflowCXVersion_update(t *testing.T) {
@@ -107,4 +110,88 @@ func testAccDialogflowCXVersion_full(context map[string]interface{}) string {
 		description  = "version 1.0.0"
 	}
 	  `, context)
+}
+
+func TestAccDialogflowCXVersion_dialogflowcxVersionFullExample(t *testing.T) {
+	t.Parallel()
+
+	context := map[string]interface{}{
+		"random_suffix": randString(t, 10),
+	}
+
+	vcrTest(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckDialogflowCXVersionDestroyProducer(t),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDialogflowCXVersion_dialogflowcxVersionFullExample(context),
+			},
+			{
+				ResourceName:            "google_dialogflow_cx_version.version_1",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"parent"},
+			},
+		},
+	})
+}
+
+func testAccDialogflowCXVersion_dialogflowcxVersionFullExample(context map[string]interface{}) string {
+	return Nprintf(`
+resource "google_dialogflow_cx_agent" "agent" {
+  display_name = "tf-test-dialogflowcx-agent%{random_suffix}"
+  location = "global"
+  default_language_code = "en"
+  supported_language_codes = ["fr","de","es"]
+  time_zone = "America/New_York"
+  description = "Example description."
+  avatar_uri = "https://cloud.google.com/_static/images/cloud/icons/favicons/onecloud/super_cloud.png"
+  enable_stackdriver_logging = true
+  enable_spell_correction    = true
+	speech_to_text_settings {
+		enable_speech_adaptation = true
+	}
+}
+
+
+resource "google_dialogflow_cx_version" "version_1" {
+  parent       = google_dialogflow_cx_agent.agent.start_flow
+  display_name = "1.0.0"
+  description  = "version 1.0.0"
+}
+`, context)
+}
+
+func testAccCheckDialogflowCXVersionDestroyProducer(t *testing.T) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		for name, rs := range s.RootModule().Resources {
+			if rs.Type != "google_dialogflow_cx_version" {
+				continue
+			}
+			if strings.HasPrefix(name, "data.") {
+				continue
+			}
+
+			config := googleProviderConfig(t)
+
+			url, err := replaceVarsForTest(config, rs, "{{DialogflowCXBasePath}}{{parent}}/versions/{{name}}")
+			if err != nil {
+				return err
+			}
+
+			billingProject := ""
+
+			if config.BillingProject != "" {
+				billingProject = config.BillingProject
+			}
+
+			_, err = sendRequest(config, "GET", billingProject, url, config.userAgent, nil)
+			if err == nil {
+				return fmt.Errorf("DialogflowCXVersion still exists at %s", url)
+			}
+		}
+
+		return nil
+	}
 }
