+++ website/docs/r/dialogflow_cx_environment.html.markdown
@@ -1,17 +1,4 @@
 ---
-# ----------------------------------------------------------------------------
-#
-#     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-#
-# ----------------------------------------------------------------------------
-#
-#     This file is automatically generated by Magic Modules and manual
-#     changes will be clobbered when the file is regenerated.
-#
-#     Please read more about how to change this file in
-#     .github/CONTRIBUTING.md.
-#
-# ----------------------------------------------------------------------------
 subcategory: "Dialogflow CX"
 page_title: "Google: google_dialogflow_cx_environment"
 description: |-
